requirements.txt
Django==5.2.6
python-decouple==3.8
django-jazzmin==3.0.1
django-test-without-migrations==0.6
model-bakery==1.20.5
sentry-sdk
pillow
django-filter
